Move all menu links into App.js file
